[UPD] Refactor sage_purchase_account and add new module choreograph_sage_sftp_import
