+++ choreograph-addons/choreograph_sage_sftp_import/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+from . import models
