+++ choreograph-addons/choreograph_sage_sftp_import/__manifest__.py
@@ -1,23 +1,20 @@
 # -*- coding: utf-8 -*-
 {
-    'name': 'Choreograph SAGE Purchase Account',
+    'name': 'Choreograph SAGE SFTP Import',
     'version': '16.0.0.0',
     'license': 'LGPL-3',
     'category': 'Accounting',
     'sequence': -98,
     'summary': 'Manage Choreograph SAGE In invoice import',
-    'description': """This module allows to import purchase invoice from sage""",
+    'description': """This module allows to import invoice from sage""",
     'author': 'ArkeUp',
     'website': 'https://arkeup.com',
     'depends': [
         'choreograph_sage_sftp_server',
-        'choreograph_sage_sftp_import',
         'choreograph_account',
         'account'
     ],
     'data': [
-        'data/ir_cron_data.xml',
-        'views/res_config_settings_views.xml'
     ],
     'assets': {
         'web._assets_primary_variables': [],
