+++ choreograph-addons/choreograph_sage_sftp_import/data/ir_cron_data.xml
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+    <data>
+        <record id="ir_cron_import_account_move_in_invoice" model="ir.cron">
+            <field name="name">Sage: Import account move in invoice</field>
+            <field name="interval_number">1</field>
+            <field name="interval_type">days</field>
+            <field name="numbercall">-1</field>
+            <field name="model_id" ref="account.model_account_move"/>
+            <field name="code">model.import_account_move_in_invoice()</field>
+            <field name="state">code</field>
+        </record>
+    </data>
+</odoo>
