+++ choreograph-addons/choreograph_sage_sftp_import/models/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+from . import account_move
